pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic_segmentation"
readme="README.md"
requires-python=">=3.7.0"
version = "0.1.0"
dependencies=[
    "protobuf",
    "numpy",
    "pytorch-lightning",
    "torch",
    "torchmetrics",
    "torchvision",
    "opencv-python",
    "Pillow",
    "tyro",
    "seaborn",
]

[tool.setuptools]
packages = ["semantic_segmentation"]

# black
[tool.black]
line-length = 120